requirements.txt
google-generativeai
gradio
Pillow